requirements.txt
Django==4.2.7
djangorestframework
psycopg2
drf-spectacular
Pillow